labs/ssvp_slt/requirements.txt
einops
evaluate
hydra-ax-sweeper
hydra-core
hydra-submitit-launcher
iopath
matplotlib
numpy
omegaconf
opencv-python-headless
pandas
psutil
pysrt
requests
rouge_score
sacrebleu
scikit-learn
sentencepiece
setuptools
simplejson 
stopes
submitit
tensorboard
timm
tqdm
transformers==4.38.0
truecase
urllib3
wandb
